requirements.txt
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-dotenv==0.19.1
Werkzeug==2.0.2
